Unify /sql/ types

+++ prisma/sql/getAllocatorBiggestClientDistribution.sql
@@ -4,11 +4,11 @@ with "allocators_with_ratio" as (select "week"                                  
                                         sum("sum_of_allocations")                             as "totalDatacap"
                                  from "client_allocator_distribution_weekly"
                                  group by "week", "allocator")
-select "week"                                                  as "week",
-       100 * ceil("biggestToTotalRatio"::float8 * 20) / 20 - 5 as "valueFromExclusive",
-       100 * ceil("biggestToTotalRatio"::float8 * 20) / 20     as "valueToInclusive",
-       count(*)::int                                           as "count",
-       sum("totalDatacap")::float                              as "totalDatacap"
+select "week"                                                 as "week",
+       100 * ceil("biggestToTotalRatio"::float * 20) / 20 - 5 as "valueFromExclusive",
+       100 * ceil("biggestToTotalRatio"::float * 20) / 20     as "valueToInclusive",
+       count(*)::int                                          as "count",
+       sum("totalDatacap")::bigint                            as "totalDatacap"
 from "allocators_with_ratio"
 group by "valueFromExclusive", "valueToInclusive", "week"
 order by "week", "valueFromExclusive";
